SimplePolymarket-Bot/requirements.txt
requests>=2.31.0
websocket-client>=1.6.4
web3>=6.11.0
py-clob-client>=0.34.5
python-dotenv>=1.0.0
